moved: new-chat and new-group into new folder
