+++ resources/views/components/actions/new-chat.blade.php
@@ -4,7 +4,7 @@
 
 
 <x-wirechat::actions.open-modal
-        component="wirechat.new-chat"
+        component="wirechat.new.chat"
         :widget="$widget"
         >
 {{$slot}}
