+++ src/Livewire/New/Group.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Namu\WireChat\Livewire;
+namespace Namu\WireChat\Livewire\New;
 
 use Livewire\Attributes\Validate;
 use Livewire\WithFileUploads;
@@ -9,7 +9,7 @@ use Namu\WireChat\Livewire\Modals\ModalComponent;
 use Namu\WireChat\Livewire\WireChat as WidgetsWireChat;
 use Namu\WireChat\Traits\Widget;
 
-class NewGroup extends ModalComponent
+class Group extends ModalComponent
 {
     use Widget;
     use WithFileUploads;
@@ -180,6 +180,6 @@ class NewGroup extends ModalComponent
     public function render()
     {
 
-        return view('wirechat::livewire.new-group', ['maxGroupMembers' => WireChat::maxGroupMembers()]);
+        return view('wirechat::livewire.new.group', ['maxGroupMembers' => WireChat::maxGroupMembers()]);
     }
 }
